CIV-4715 refactor place in the correct directory
